Remove libsweep from repo; Update Dockerfile to install sweep-sdk
The catkin-ized version of libsweep was being properly built by catkin, but not being correctly referenced by sweep-ros for the required build time dependencies. This is easy enough to install via Dockerfile in any case -- perhaps not as painless for non-Docker users, but at least this will guarantee the expected build behavior upon Docker container generation.